Load menu improvements
- Centered the in-game load menu.
- Moved save file loading logic to Civ3FileDialog to reduce duplication.
- Removed an unclear and likely outdated comment from Civ3FileDialog.
- The popup menu now disappears after opening the load menu, similar to Civ3.
