+++ C7/UIElements/Civ3FileDialog.cs
@@ -1,19 +1,30 @@
 using Godot;
+using Serilog;
 
 [GlobalClass]
 public partial class Civ3FileDialog : FileDialog {
-	// Use this instead of a scene-based FileDialog to avoid it saving the local dev's last browsed folder in the repo
-	// While instantiated it will return to the last-accessed folder when reopened
-	public Civ3FileDialog() {
-		FileMode = FileDialog.FileModeEnum.OpenFile;
-	}
+	// An object for passing information (like save file paths) between scenes.
+	GlobalSingleton Global;
+	private ILogger log;
 
 	public override void _Ready() {
-		Access = AccessEnum.Filesystem;
 		base._Ready();
+		log = LogManager.ForContext<Civ3FileDialog>();
+
+		FileMode = FileDialog.FileModeEnum.OpenFile;
+		Access = AccessEnum.Filesystem;
+
+		Global = GetNode<GlobalSingleton>("/root/GlobalSingleton");
+		FileSelected += OnFileSelected;
 	}
 
 	public void SetDirectory(string RelPath) {
 		CurrentDir = Util.Civ3Root + "/" + RelPath;
+	}
+
+	private void OnFileSelected(string path) {
+		log.Information($"loading {path}");
+		Global.LoadGamePath = path;
+		GetTree().ChangeSceneToFile("res://C7Game.tscn");
 	}
 }
