+++ C7/UIElements/Popups/GameMenu.cs
@@ -3,13 +3,6 @@ using Godot;
 using Serilog;
 
 public partial class GameMenu : Popup {
-	private ILogger log;
-
-	Civ3FileDialog GameMenuLoadDialog;
-
-	// An object for passing information (like save file paths) between scenes.
-	GlobalSingleton Global;
-
 	public GameMenu() {
 		alignment = BoxContainer.AlignmentMode.Center;
 		margins = new Margins(top: 100);
@@ -17,7 +10,6 @@ public partial class GameMenu : Popup {
 
 	public override void _Ready() {
 		base._Ready();
-		log = LogManager.ForContext<GameMenu>();
 
 		AddTexture(370, 300);
 		AddBackground(370, 300);
@@ -32,12 +24,6 @@ public partial class GameMenu : Popup {
 		AddButton("Save Game (Ctrl-S)", 185, save);
 		AddButton("Quit Game (ESC)", 210, quit);
 
-		Global = GetNode<GlobalSingleton>("/root/GlobalSingleton");
-
-		GameMenuLoadDialog = new Civ3FileDialog();
-		AddChild(GameMenuLoadDialog);
-		GameMenuLoadDialog.SetDirectory(@"Conquests/Saves");
-		GameMenuLoadDialog.FileSelected += OnFileSelected;
 	}
 
 	private void save() {
@@ -53,11 +39,15 @@ public partial class GameMenu : Popup {
 	}
 
 	private void load() {
-		log.Information("load game button pressed");
+		var loadDialog = GetNode<Civ3FileDialog>("../%LoadDialog");
+		loadDialog.SetDirectory(@"Conquests/Saves");
+
 		// TODO: The main menu does sound playing but we don't know our path in
 		// the scene, which makes this hard.
 		// PlayButtonPressedSound();
-		GameMenuLoadDialog.Popup();
+		GetParent().EmitSignal(PopupOverlay.SignalName.HidePopup);
+
+		loadDialog.Popup();
 	}
 
 	private void quit() {
@@ -70,11 +60,5 @@ public partial class GameMenu : Popup {
 
 	private void map() {
 		GetParent().EmitSignal(PopupOverlay.SignalName.HidePopup);
-	}
-
-	private void OnFileSelected(string path) {
-		log.Information($"loading {path}");
-		Global.LoadGamePath = path;
-		GetTree().ChangeSceneToFile("res://C7Game.tscn");
 	}
 }
